+++ src/createGlobalState.ts
@@ -11,8 +11,8 @@ type ExportFields =
  *
  * It returns a set of functions
  * - `useGlobalState`: a custom hook works like React.useState
- * - `getGlobalState`: a function to get a global state by name outside React
- * - `setGlobalState`: a function to set a global state by name outside React
+ * - `getGlobalState`: a function to get a global state by key outside React
+ * - `setGlobalState`: a function to set a global state by key outside React
  *
  * @example
  * import { createGlobalState } from 'react-hooks-global-state';